+++ tests/testthat/test-gather_ard.R
@@ -1,5 +1,5 @@
 skip_on_cran()
-skip_if_not(is_pkg_installed(c("cardx", "broom.helpers", "car", "parameters")))
+skip_if_not(is_pkg_installed(c("broom.helpers", "car", "parameters")))
 
 test_that("gather_ard(x) works with `tbl_*()` functions", {
   # tbl_summary()