+++ tests/testthat/test-modify_caption.R
@@ -12,7 +12,7 @@ test_that("modify_caption(caption) works", {
 })
 
 test_that("modify_caption() works with tbl_svysummary()", {
-  skip_if_not(is_pkg_installed(c("survey", "cardx")))
+  skip_if_not(is_pkg_installed("survey"))
 
   expect_equal(
     survey::svydesign(~1, data = as.data.frame(Titanic), weights = ~Freq) |>
@@ -49,7 +49,7 @@ test_that("modify_caption() works with tbl_cross()", {
 })
 
 test_that("modify_caption() works with tbl_regression()", {
-  skip_if_not(is_pkg_installed(c("cardx", "broom.helpers")))
+  skip_if_not(is_pkg_installed("broom.helpers"))
 
   expect_equal(glm(response ~ age + grade, trial, family = binomial()) |>
                  tbl_regression(exponentiate = TRUE) |>
@@ -62,7 +62,7 @@ test_that("modify_caption() works with tbl_regression()", {
 })
 
 test_that("modify_caption() works with tbl_uvregression()", {
-  skip_if_not(is_pkg_installed(c("cardx", "broom.helpers")))
+  skip_if_not(is_pkg_installed("broom.helpers"))
   expect_equal(tbl_uvregression(trial, method = glm, y = response, method.args = list(family = binomial),
                                 exponentiate = TRUE, include = c("age", "grade")) |>
                  modify_caption("**Adding a caption**", text_interpret = "html") |>