+++ tests/testthat/test-modify_header.R
@@ -1,5 +1,5 @@
 skip_on_cran()
-skip_if_not(is_pkg_installed(c("cardx", "broom.helpers")))
+skip_if_not(is_pkg_installed("broom.helpers"))
 
 # first, testing deprecation
 test_that("modify_header(update,quiet) are deprecated", {
@@ -152,7 +152,7 @@ test_that("modify_header(text_interpret) works", {
 })
 
 test_that("modify_header() works with tbl_svysummary()", {
-  skip_if_not(is_pkg_installed(c("survey", "cardx")))
+  skip_if_not(is_pkg_installed("survey"))
 
   expect_equal(
     survey::svydesign(~1, data = as.data.frame(Titanic), weights = ~Freq) |>